feat: improved caching and shit
