+++ config.yaml
@@ -1,3 +1,3 @@
 repositories:
-  prometheus/prometheus: ^2.1.0
-  caarlos0/version_exporter: 0.0.5
+  prometheus/prometheus: 2.5.0
+  caarlos0/version_exporter: 1.0.2
